use one global instead of duplicating API URL
Quick note on stand model's URL generation...

1. This seems to work: having a `urlRoot` property set to `http://lestands-api.herokuapp.com/stands`. By setting `urlRoot`, it tells backbone to construct the URL for the specific item as `urlRoot` + `/:id`. The documentation implies that urlRoot and url() aren't strictly necessary (http://backbonejs.org/#Model-url "Generates URLs of the form: "[collection.url]/[id]" by default") but I find that without at least one of those present, I get the following error in the console: " Uncaught Error: A "url" property or function must be specified".

2. So we don't duplicate the url, I'm creating a global in app.js: `var API_BASE = "https://lestands-api.herokuapp.com"`. Then we can simplify `urlRoot` by not needing to have the URL in both the stand.js model file and stands.js collection, as we'll just set `urlRoot` and `url` (respectively) to `API_BASE + '/stands'`.

3. The alternative to supplying `urlRoot` in the stand.js model is to provide the `url()` method which constructs the URL for any given stand. It'd look like this:

```
url: function() {
        return _.result(app.stands, 'url') + '/' + this.id;
     },
```

That function calls the `url()` method on the app.stands collection (using underscore's neat `result` method that fetches either a property value or a function return value depending on the type of the property you pass in, which is a method in our case. http://underscorejs.org/#result). The url is "http://lestands-api.herokuapp.com", to which "/:id" is appended.  On the one hand, I like that url() spells out what the url of the model is, but on the other hand, it is cleaner perhaps to just provide the urlRoot, so I'll probably stick to that.